refactor into one class
