+++ src/Nethermind/Nethermind.State/Proofs/PatriciaTrieT.cs
@@ -7,10 +7,9 @@ using System.Linq;
 using Nethermind.Db;
 using Nethermind.Logging;
 using Nethermind.Serialization.Rlp;
-using Nethermind.State.Proofs;
 using Nethermind.Trie;
 
-namespace Nethermind.State.Trie;
+namespace Nethermind.State.Proofs;
 
 /// <summary>
 /// An abstract class that represents a Patricia trie built of a collection of <see cref="T"/>.
