+++ src/Nethermind/Nethermind.State/Proofs/WithdrawalTrie.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Generic;
 using Nethermind.Core;
 using Nethermind.Serialization.Rlp;
-using Nethermind.State.Trie;
 
 namespace Nethermind.State.Proofs;
